added function to save user-generated haiku in app
wrote saveUserHaiku method that works.  Still need to test whether save
to plist (added in viewDidUnload) works.

Futzed with what happens if user goes through all haiku and realized
that there's a problem--self.theseAreDone and self.indx increase
irrespective of the category (my haiku, user haiku).  We need to keep
track of which category the next haiku is in.  I can easily do it
through brute force but finding an elegant solution will be slightly
trickier.