cleaned things up, failed to fix bugs
still to do:

1.  Write code that runs haikuInstructions if user presses "My Haiku"
in the segmented controller and it's never been pressed before.

2.  Get "back" button to appear off user clicks a link in UIWebView.
Is it necessary to create GHWebView as a separate class?

3.  Write code that gives error if user tries to click on Amazon and a
connection isn't available.

4.  Give user chance to opt out of sending haiku s/he composes to
central database.

5.  Connect app to central database so that haiku can be sent there.

6.  Write method to check haiku user writes against haiku already
written so there are no duplicates?

7.  Write method allowing user to delete haiku s/he's written.

8.  Question:  how will it affect the user's experience if/when haiku
s/he's already seen in "user" or "Derfner" categories reappear in "all"
category?  Will this need to be adjusted?  If so, how?

9.  Need to test to make sure it starts over once all 110 haiku have
been seen.