Added AlertView, dealt with bug
There's now an alert when user clicks on a link in UIWebView and
there's no Internet connection.

The bug that made the keyboard pop up when the user touched the
lower-left-hand side of the screen is fixed.

Still need to add back and forward buttons to UIWebView and connect app
to database.